+++ OpenSolarMax.Game/Modding/Modding.cs
@@ -1,112 +1,15 @@
 using System.Collections.Immutable;
 using System.Reflection;
-using System.Runtime.Loader;
-using CsToml.Extensions.Configuration;
-using Microsoft.Extensions.Configuration;
 using Myra.Graphics2D.UI;
-using Nine.Assets;
 using OpenSolarMax.Game.Modding.Concept;
 using OpenSolarMax.Game.Modding.Declaration;
 using OpenSolarMax.Game.Modding.ECS;
 using OpenSolarMax.Game.Modding.UI;
-using Zio;
-using Zio.FileSystems;
 
 namespace OpenSolarMax.Game.Modding;
 
 internal static partial class Modding
 {
-    #region 工厂方法
-
-    internal static BehaviorMod LoadBehaviorMod(
-        BehaviorModInfo info,
-        IReadOnlyDictionary<string, Assembly> sharedAssemblies
-    )
-    {
-        // 加载程序集
-        var ctx = new ModLoadContext(info.Assembly, sharedAssemblies);
-        using var dllStream = info.Assembly.Open(FileMode.Open, FileAccess.Read);
-#if DEBUG
-        var pdb = info
-            .Assembly.Directory.EnumerateFiles($"{info.Assembly.NameWithoutExtension}.pdb")
-            .FirstOrDefault();
-        using var pdbStream = pdb?.Open(FileMode.Open, FileAccess.Read);
-        var assembly = ctx.LoadFromStream(dllStream, pdbStream);
-#else
-        var assembly = ctx.LoadFromStream(dllStream);
-#endif
-
-        // 加载资产文件系统
-        List<IFileSystem> contentFileSystems = [new ResourceFileSystem(assembly)];
-        if (info.Content is not null)
-        {
-            contentFileSystems.Add(
-                new SubFileSystem(info.Content.FileSystem, info.Content.Path, owned: false)
-            );
-        }
-
-        // 加载配置文件
-        IConfigurationRoot? configs = null;
-        if (info.Configs is not null)
-        {
-            var configsBuilder = new ConfigurationBuilder();
-            using var tomlStream = info.Configs.Open(
-                FileMode.Open,
-                FileAccess.Read,
-                FileShare.Read
-            );
-            configsBuilder.AddTomlStream(tomlStream);
-            configs = configsBuilder.Build();
-        }
-
-        return new BehaviorMod
-        {
-            Metadata = info,
-            Assembly = assembly,
-            ContentFileSystems = contentFileSystems.ToImmutableArray(),
-            Configs = configs,
-            // 查找组件类型
-            ComponentTypes = assembly
-                .ExportedTypes.Where(t => t.GetCustomAttribute<ComponentAttribute>() is not null)
-                .ToImmutableArray(),
-            // 查找关卡文件声明类型
-            DeclarationSchemaInfos = FindDeclarationTypes(assembly).ToImmutableDictionary(),
-            // 查找游玩场景行为相关类型
-            GameplayBehaviorsInfo = new BehaviorsInfo(
-                FindTranslatorTypes(assembly, GameplayOrPreview.Gameplay).ToImmutableDictionary(),
-                FindConceptRelatedTypes(assembly, GameplayOrPreview.Gameplay)
-                    .ToImmutableDictionary(),
-                FindSystemTypes(assembly, GameplayOrPreview.Gameplay),
-                FindHookImplementations(assembly, GameplayOrPreview.Gameplay)
-                    .ToImmutableDictionary(g => g.Key, g => g.ToImmutableArray())
-            ),
-            // 查找预览场景行为相关类型
-            PreviewBehaviorsInfo = new BehaviorsInfo(
-                FindTranslatorTypes(assembly, GameplayOrPreview.Preview).ToImmutableDictionary(),
-                FindConceptRelatedTypes(assembly, GameplayOrPreview.Preview)
-                    .ToImmutableDictionary(),
-                FindSystemTypes(assembly, GameplayOrPreview.Preview),
-                FindHookImplementations(assembly, GameplayOrPreview.Preview)
-                    .ToImmutableDictionary(g => g.Key, g => g.ToImmutableArray())
-            ),
-        };
-    }
-
-    internal static ContentMod LoadContentMod(ContentModInfo info)
-    {
-        return new ContentMod
-        {
-            Metadata = info,
-            // 加载资产文件系统
-            ContentFileSystems =
-            [
-                new SubFileSystem(info.Content.FileSystem, info.Content.Path, owned: false),
-            ],
-        };
-    }
-
-    #endregion
-
     #region 反射扫描
 
     /// <summary>
